Moved column adjusting funtionality to parent component
