+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/visualization/results-visual/table-header.tsx
@@ -35,6 +35,9 @@ export type Header = {
 
 type HeaderProps = {
   lazyResults: LazyQueryResults
+  columns: Array<any>
+  mouseDownMain: Function
+  attributes: any
 }
 
 type Sort = {
@@ -159,82 +162,82 @@ export const HeaderCheckbox = ({
   )
 }
 
-export const Header = ({ lazyResults }: HeaderProps) => {
+export const Header = ({ lazyResults, columns, mouseDownMain,attributes }: HeaderProps) => {
   const handleSortClick = _.debounce(updateSort, 500, true)
-  const [shownAttributes, setShownAttributes] = React.useState(
-    TypedUserInstance.getResultsAttributesShownTable()
-  )
-  const [activeIndex, setActiveIndex] = React.useState(null);
-  const minCellWidth = 200
+  // const [shownAttributes, setShownAttributes] = React.useState(
+  //   TypedUserInstance.getResultsAttributesShownTable()
+  // )
+  // const { listenTo } = useBackbone()
+
+  // const [activeIndex, setActiveIndex] = React.useState(activeIndexMain)
+  // const minCellWidth = 200
   
-  const createHeaders = () => {
-    return shownAttributes.map((item) => ({
-      attr: item,
-      ref: React.useRef<HTMLDivElement>(null)
-    }))
-  }
+  // const createHeaders = () => {
+  //   return shownAttributes.map((item) => ({
+  //     attr: item,
+  //     ref: React.useRef<HTMLDivElement>(null)
+  //   }))
+  // }
 
-  const columns = createHeaders()
+  // const columns = createHeaders()
 
-  const mouseDown = (index:any) => {
-    setActiveIndex(index);
-  }
+  // const mouseDown = (index:any) => {
+  //   setActiveIndex(index);
+  // }
 
-  const { listenTo } = useBackbone()
+  // const mouseMove = React.useCallback((e) => {
 
-  const mouseMove = React.useCallback((e) => {
-
-    columns.map((col, i) => {
-      if (i === activeIndex) {
-        if (col.ref.current){
-          const width = e.clientX - col.ref.current?.getBoundingClientRect().x
-          if (width > minCellWidth){
-            col.ref.current.style.width = `${width}px`
-          }
-        }   
-      }
+  //   columns.map((col, i) => {
+  //     if (i === activeIndex) {
+  //       if (col.ref.current){
+  //         const width = e.clientX - col.ref.current?.getBoundingClientRect().x
+  //         if (width > minCellWidth){
+  //           col.ref.current.style.width = `${width}px`
+  //         }
+  //       }   
+  //     }
   
-    });
+  //   });
   
-  }, [activeIndex, columns])
+  // }, [activeIndex, columns])
 
-  const removeListeners = React.useCallback(() => {
-    window.removeEventListener('mousemove', mouseMove)
-    window.removeEventListener('mouseup', removeListeners)
-  }, [mouseMove])
+  // const removeListeners = React.useCallback(() => {
+  //   window.removeEventListener('mousemove', mouseMove)
+  //   window.removeEventListener('mouseup', removeListeners)
+  // }, [mouseMove])
   
-  const mouseUp = React.useCallback(() => {
-    setActiveIndex(null)
-    removeListeners()
-  }, [setActiveIndex, removeListeners])
+  // const mouseUp = React.useCallback(() => {
+  //   setActiveIndex(null)
+  //   removeListeners()
+  // }, [setActiveIndex, removeListeners])
 
-  React.useEffect(() => {
-    if (activeIndex !== null) {
-      window.addEventListener('mousemove', mouseMove)
-      window.addEventListener('mouseup', mouseUp)
-    }
+  // React.useEffect(() => {
+  //   if (activeIndex !== null) {
+  //     window.addEventListener('mousemove', mouseMove)
+  //     window.addEventListener('mouseup', mouseUp)
+  //   }
   
-    return () => {
-      removeListeners()
-    }
-  }, [activeIndex, mouseMove, mouseUp, removeListeners])
+  //   return () => {
+  //     removeListeners()
+  //   }
+  // }, [activeIndex, mouseMove, mouseUp, removeListeners])
 
-  React.useEffect(() => {
-    listenTo(
-      user.get('user').get('preferences'),
-      'change:results-attributesShownTable',
-      () => {
-        setShownAttributes(TypedUserInstance.getResultsAttributesShownTable())
-      }
-    )
-  }, [])
+  // React.useEffect(() => {
+  //   listenTo(
+  //     user.get('user').get('preferences'),
+  //     'change:results-attributesShownTable',
+  //     () => {
+  //       setShownAttributes(TypedUserInstance.getResultsAttributesShownTable())
+  //     }
+  //   )
+  // }, [])
   return (
     <React.Fragment>
       <div
         data-id="table-container"
         className="bg-inherit whitespace-no-wrap flex items-strech flex-no-wrap"
         style={{
-          width: shownAttributes.length * 200 + 'px',
+          width: attributes.length * 200 + 'px',
           display: 'grid',
           gridTemplateColumns: 'repeat(12, 1fr)'
         }}
@@ -255,13 +258,17 @@ export const Header = ({ lazyResults }: HeaderProps) => {
               sortable ? 'is-sortable' : ''
             } Mui-border-divider border border-t-0 border-l-0 border-b-0`}
             style={{cursor: 'col-resize', display: 'flex'}}
-            onMouseDown={() => mouseDown(index)}
+            onMouseDown={() => {
+              // mouseDown(index)
+              mouseDownMain(index)
+            }}
             >
                 <CellComponent
                   data-propertyid={`${attr}`}
                   data-propertytext={`${label ? `${label}` : `${attr}`}`}
                   style={{
-                    minWidth: '200px',
+                    width: '100%',
+                    minWidth: '200px'
                   }}
                 >                    
                     <Button
@@ -279,7 +286,7 @@ export const Header = ({ lazyResults }: HeaderProps) => {
                         </span>
                         <span
                           className={getSortDirectionClass(attr)}
-                          style={{ paddingLeft: '3px' }}
+                          style={{ paddingLeft: '3px'}}
                         />
                       </div>
                     </Button>
